File Structure and Addition Pages
Created forms and pages for additions of schools, reviews, and professors.
Added a logout page that redirects to the login after 5 seconds using a javascript script.
Moved the css file to the css folder in public and updated link in main.handlebars.
Created routes to the new addition pages and put their links on the login page (for test)